Test refactoring.
